# apps/api/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: PricePulse API
  version: 1.0.0
  description: REST API for tracking simulated e-commerce product prices.
servers:
  - url: http://localhost:4000
paths:
  /health:
    get:
      summary: Check API health
      responses:
        "200":
          description: API is healthy
  /api/products:
    get:
      summary: List tracked products
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
            maximum: 50
        - in: query
          name: trend
          schema:
            type: string
            enum: [up, down, stable]
      responses:
        "200":
          description: Paginated product list
    post:
      summary: Track a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProduct"
      responses:
        "201":
          description: Product created
        "400":
          description: Invalid input
  /api/products/{id}:
    delete:
      summary: Stop tracking a product
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Product deleted
        "404":
          description: Product not found
  /api/products/{id}/price:
    patch:
      summary: Update a product price from the background worker
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPrice]
              properties:
                currentPrice:
                  type: number
                  minimum: 0.01
      responses:
        "200":
          description: Product updated
        "400":
          description: Invalid input
        "404":
          description: Product not found
  /api/products/{id}/simulate:
    post:
      summary: Simulate one price update for a product
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Product updated
        "404":
          description: Product not found
components:
  schemas:
    CreateProduct:
      type: object
      required: [name, url, currentPrice]
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 120
        url:
          type: string
          format: uri
        currentPrice:
          type: number
          minimum: 0.01